Create Lift method OperateLift. Add first part of code to it.
The OperateLift method will control the lifts behaviour and how it responds to lift calls, and proceeds to travel to selected destination floors. Code added that updates the selected and called  floors, the  lift riders,  the lift callers sets when the lift is at a stop; all necessary for determining the lifts next stop. Code added to also record all journey details for the lift and for the callers.