Create Lift method IncorporateCalledFloors
Method responsible for incorporating called floors into the lift path. It determines which called floors are being passed-by along the lift path for selected floors; then picks a called floor for which the caller is travelling in the right direction; if no floors exist, then all other called floors are trreated as selected floors and the route is then optimised.